fix indentation

+++ Makefile
@@ -305,39 +305,39 @@ ifeq ($$(strip $$(DL_SITE$(1))),)
 else
 	@if [ "$(FREETZ_DL_DETECT_IMAGE_NAME)" == "y" ]; then $(RM) -f "$(DL_FW_DIR)/$(DL_SOURCE)"; fi; \
 	if [ ! -e "$$(DL_FW_DIR)/$$(DL_SOURCE$(1))" ]; then \
-	if [ -n "$$(DL_SOURCE$(1)_CONTAINER)" ]; then \
-		if [ ! -r $$(DL_FW_DIR)/$$(DL_SOURCE$(1)_CONTAINER) ]; then \
-			if ! $$(DL_TOOL) --no-append-servers $$(DL_FW_DIR) "$$(DL_SOURCE$(1)_CONTAINER)" "$$(DL_SITE$(1))" $$(DL_SOURCE$(1)_CONTAINER_MD5) $$(SILENT); then \
-				$$(call ERROR,3,Could not download firmware image. See https://freetz.github.io/wiki/FAQ#Couldnotdownloadfirmwareimage for details.) \
+		if [ -n "$$(DL_SOURCE$(1)_CONTAINER)" ]; then \
+			if [ ! -r $$(DL_FW_DIR)/$$(DL_SOURCE$(1)_CONTAINER) ]; then \
+				if ! $$(DL_TOOL) --no-append-servers $$(DL_FW_DIR) "$$(DL_SOURCE$(1)_CONTAINER)" "$$(DL_SITE$(1))" $$(DL_SOURCE$(1)_CONTAINER_MD5) $$(SILENT); then \
+					$$(call ERROR,3,Could not download firmware image. See https://freetz.github.io/wiki/FAQ#Couldnotdownloadfirmwareimage for details.) \
+				fi; \
 			fi; \
-		fi; \
-		case "$$(DL_SOURCE$(1)_CONTAINER_SUFFIX)" in \
-			.zip|.ZIP) \
-				if [ "$$(FREETZ_DL_DETECT_IMAGE_NAME)" == "y" ]; then \
-					DL_SOURCE_DETECTED=$$$$(unzip -j $$(QUIETSHORT) -l $$(DL_FW_DIR)/$$(DL_SOURCE$(1)_CONTAINER) *.image | sed -rn 's/.*( |\/)(.*\.image)/\2/p'); \
-					echo "Using detected .image name: $$$$DL_SOURCE_DETECTED" >/dev/null; \
-				else \
-					DL_SOURCE_DETECTED=$$(DL_SOURCE); \
-					echo "Using hardcoded .image name: $$$$DL_SOURCE_DETECTED" >/dev/null; \
-				fi; \
-				if [ ! -f $$(DL_FW_DIR)/$$$$DL_SOURCE_DETECTED ]; then \
-					echo "Unzipping archive file: $$(DL_SOURCE$(1)_CONTAINER)" >/dev/null; \
-					if ! unzip -j $$(QUIETSHORT) $$(DL_FW_DIR)/$$(DL_SOURCE$(1)_CONTAINER) *$$$$DL_SOURCE_DETECTED -d $$(DL_FW_DIR); then \
-						$$(call ERROR,3,Could not unzip firmware image.) \
+			case "$$(DL_SOURCE$(1)_CONTAINER_SUFFIX)" in \
+				.zip|.ZIP) \
+					if [ "$$(FREETZ_DL_DETECT_IMAGE_NAME)" == "y" ]; then \
+						DL_SOURCE_DETECTED=$$$$(unzip -j $$(QUIETSHORT) -l $$(DL_FW_DIR)/$$(DL_SOURCE$(1)_CONTAINER) *.image | sed -rn 's/.*( |\/)(.*\.image)/\2/p'); \
+						echo "Using detected .image name: $$$$DL_SOURCE_DETECTED" >/dev/null; \
+					else \
+						DL_SOURCE_DETECTED=$$(DL_SOURCE); \
+						echo "Using hardcoded .image name: $$$$DL_SOURCE_DETECTED" >/dev/null; \
 					fi; \
-				fi; \
-				if [ "$$(FREETZ_DL_DETECT_IMAGE_NAME)" == "y" ]; then \
-					[ -f $$(DL_FW_DIR)/$$$$DL_SOURCE_DETECTED ] && ln $$(DL_FW_DIR)/$$$$DL_SOURCE_DETECTED $$(DL_FW_DIR)/$$(DL_SOURCE$(1)); \
-					echo "Created hardlink for .image file: $(DL_SOURCE)" >/dev/null; \
-				fi; \
-				;; \
-			*) \
-				$$(call ERROR,3,Not able to extract $$(DL_SOURCE_CONTAINER_SUFFIX) firmware archives at all.) \
-				;; \
-		esac \
-	elif ! $$(DL_TOOL) --no-append-servers $$(DL_FW_DIR) "$$(DL_SOURCE$(1))" "$$(DL_SITE$(1))" $$(DL_SOURCE$(1)_MD5) $$(SILENT); then \
-		$$(call ERROR,3,Could not download firmware image. See https://freetz.github.io/wiki/FAQ#Couldnotdownloadfirmwareimage for details.) \
-	fi; \
+					if [ ! -f $$(DL_FW_DIR)/$$$$DL_SOURCE_DETECTED ]; then \
+						echo "Unzipping archive file: $$(DL_SOURCE$(1)_CONTAINER)" >/dev/null; \
+						if ! unzip -j $$(QUIETSHORT) $$(DL_FW_DIR)/$$(DL_SOURCE$(1)_CONTAINER) *$$$$DL_SOURCE_DETECTED -d $$(DL_FW_DIR); then \
+							$$(call ERROR,3,Could not unzip firmware image.) \
+						fi; \
+					fi; \
+					if [ "$$(FREETZ_DL_DETECT_IMAGE_NAME)" == "y" ]; then \
+						[ -f $$(DL_FW_DIR)/$$$$DL_SOURCE_DETECTED ] && ln $$(DL_FW_DIR)/$$$$DL_SOURCE_DETECTED $$(DL_FW_DIR)/$$(DL_SOURCE$(1)); \
+						echo "Created hardlink for .image file: $(DL_SOURCE)" >/dev/null; \
+					fi; \
+					;; \
+				*) \
+					$$(call ERROR,3,Not able to extract $$(DL_SOURCE_CONTAINER_SUFFIX) firmware archives at all.) \
+					;; \
+			esac \
+		elif ! $$(DL_TOOL) --no-append-servers $$(DL_FW_DIR) "$$(DL_SOURCE$(1))" "$$(DL_SITE$(1))" $$(DL_SOURCE$(1)_MD5) $$(SILENT); then \
+			$$(call ERROR,3,Could not download firmware image. See https://freetz.github.io/wiki/FAQ#Couldnotdownloadfirmwareimage for details.) \
+		fi; \
 	fi
 endif
 endif
